Using hand-built configurations. Removed anansi
