--- .browserslistrc
@@ -1,1 +1,0 @@
-extends @anansi/browserslist-config
