--- .storybook/.babelrc.js
@@ -1,3 +1,0 @@
-module.exports = {
-  presets: [['@anansi/babel-preset', { typing: 'typescript', loose:true }]],
-};
