--- .storybook/main.js
@@ -1,19 +1,0 @@
-module.exports = {
-  core: {
-    builder: "webpack5",
-  },
-  stories: [
-    '../src/*.stories.@(js|jsx|ts|tsx)',
-    '../src/**/*.stories.@(js|jsx|ts|tsx)'
-  ],
-  addons: [
-    '@storybook/addon-links',
-    '@storybook/addon-essentials'
-  ],
-  typescript: {
-    reactDocgen: 'react-docgen-typescript',
-  },
-  reactOptions: {
-    fastRefresh: true,
-  },
-};
