--- .storybook/webpack.config.js
@@ -1,4 +1,0 @@
-
-const { makeStorybookConfigGenerator } = require('@anansi/webpack-config');
-
-module.exports = makeStorybookConfigGenerator(require('../webpack.config'));
