+++ package.json
@@ -1,123 +1,27 @@
 {
   "name": "@react-noui/create-form",
-  "version": "1.0.1",
+  "version": "1.0.2",
+  "description": "React form context/provider",
+  "main": "lib/index.js",
+  "author": "taystack",
+  "license": "MIT",
   "packageManager": "yarn@3.1.1",
-  "description": "react-create-form - Simple, React Forms API",
   "scripts": {
-    "lint": "eslint src --ext .ts,.tsx",
-    "format": "yarn run lint --fix",
-    "test:type": "tsc",
-    "test": "NODE_ENV=test jest",
-    "test:ci": "yarn test -- --ci",
-    "test:watch": "yarn test -- --watch",
-    "test:coverage": "yarn test -- --coverage",
-    "build:lib": "NODE_ENV=production babel src --out-dir lib --source-maps inline --extensions '.ts,.tsx,.js' --ignore '**/__tests__/**' --ignore '**/stories/**' --ignore '**/*.d.ts'",
-    "build:types": "tsc --project tsconfig.lib.json",
-    "build:clean": "rimraf lib && rimraf dist && rimraf *.tsbuildinfo",
-    "build": "webpack --mode=production",
-    "dev": "yarn run build:lib -w",
-    "prepack": "yarn run build",
-    "storybook": "start-storybook -p 6006",
-    "build:storybook": "build-storybook",
-    "start": "webpack serve --mode=development",
-    "start:server": "serve dist",
-    "build:analyze": "webpack --mode=production --env analyze",
-    "build:profile": "webpack --mode=production --env profile",
-    "test:pkg": "webpack --env check=nobuild",
-    "postinstall": "rm -rf node_modules/.cache"
+    "build": "yarn build:clean; yarn build:types; yarn build:lib",
+    "build:types": "yarn tsc",
+    "build:lib": "NODE_ENV=production yarn babel src --out-dir lib --source-maps inline --extensions '.ts,.tsx,.js' --ignore '**/__tests__/**' --ignore '**/stories/**'",
+    "build:clean": "rimraf lib"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+ssh://git@github.com/react-noui/create-form.git"
-  },
-  "bugs": {
-    "url": "https://github.com/react-noui/create-form/issues"
-  },
-  "homepage": "https://github.com/react-noui/create-form#readme",
-  "keywords": [
-    "anansi"
-  ],
-  "sideEffects": false,
-  "main": "dist/index.cjs.js",
-  "module": "lib/index.js",
-  "unpkg": "dist/index.umd.min.js",
-  "types": "lib/index.d.ts",
-  "files": [
-    "src",
-    "lib",
-    "LICENSE",
-    "README.md"
-  ],
-  "resolutions": {
-    "@types/webpack": "^5.0.0",
-    "immer": "^9.0.0",
-    "webpack": "^5.73.0",
-    "css-loader": "^6.0.0",
-    "dotenv-webpack": "^8.0.0",
-    "html-webpack-plugin": "^5.0.0",
-    "style-loader": "^3.0.0",
-    "terser-webpack-plugin": "^5.2.1",
-    "webpack-virtual-modules": "^0.4.2",
-    "react-error-overlay": "6.0.9"
-  },
-  "license": "Apache-2.0",
   "devDependencies": {
-    "@anansi/babel-preset": "3.4.1",
-    "@anansi/browserslist-config": "1.4.1",
-    "@anansi/eslint-plugin": "0.15.3",
-    "@anansi/jest-preset": "0.8.2",
-    "@anansi/webpack-config": "14.1.5",
-    "@babel/cli": "7.20.7",
-    "@babel/core": "7.20.12",
-    "@linaria/babel-preset": "4.4.1",
-    "@linaria/core": "4.2.6",
-    "@linaria/react": "4.3.4",
-    "@linaria/shaker": "4.2.8",
-    "@storybook/addon-essentials": "6.5.16",
-    "@storybook/addon-links": "6.5.16",
-    "@storybook/addons": "6.5.16",
-    "@storybook/builder-webpack5": "6.5.16",
-    "@storybook/manager-webpack5": "6.5.16",
-    "@storybook/react": "6.5.16",
-    "@testing-library/react": "^14.0.0",
-    "@testing-library/react-hooks": "^8.0.1",
-    "@types/jest": "29.4.0",
-    "@types/react": "18.0.28",
-    "@types/react-dom": "18.0.10",
-    "@types/react-test-renderer": "18.0.0",
-    "@typescript-eslint/eslint-plugin": "^5.11.0",
-    "@zerollup/ts-transform-paths": "1.7.18",
-    "eslint": "^7.0.0 || ^8.0.0",
-    "eslint-plugin-import": "^2.22.0",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.22.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "jest": "29.4.2",
-    "jest-environment-jsdom": "29.4.2",
-    "prettier": "^2.2.1",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-test-renderer": "18.2.0",
-    "rimraf": "4.1.2",
-    "rollup": "3.15.0",
-    "rollup-plugin-babel": "4.4.0",
-    "rollup-plugin-commonjs": "10.1.0",
-    "rollup-plugin-filesize": "9.1.2",
-    "rollup-plugin-json": "4.0.0",
-    "rollup-plugin-node-resolve": "5.2.0",
-    "rollup-plugin-replace": "2.2.0",
-    "rollup-plugin-terser": "7.0.2",
-    "serve": "14.2.0",
-    "typescript": "4.9.5",
-    "webpack": "5.75.0",
-    "webpack-cli": "5.0.1",
-    "webpack-dev-server": "4.11.1"
+    "@babel/cli": "^7.21.5",
+    "@babel/core": "^7.21.8",
+    "@babel/preset-env": "^7.21.5",
+    "@babel/preset-react": "^7.18.6",
+    "@babel/preset-typescript": "^7.21.5",
+    "@types/react": "^18.2.6",
+    "typescript": "^5.0.4"
   },
   "dependencies": {
-    "@babel/runtime": "7.20.13"
-  },
-  "peerDependencies": {
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0"
+    "react": "^18.2.0"
   }
 }
