--- src/__tests__/makeOptions.test.ts
@@ -1,10 +1,0 @@
-import { makeOptions } from 'makeOptions';
-
-describe('makeOptions', () => {
-  test('defaults to controlled inputs', () => {
-    expect(makeOptions()).toEqual({
-      validate: {},
-      props: {},
-    });
-  });
-});
