+++ src/index.ts
@@ -1,41 +1,4 @@
 export * from './createForm';
 export * from './createFormUncontrolled';
-export * from './hooks/controlled';
-export * from './hooks/uncontrolled';
-export type {
-  Primitive,
-  PrimitiveRecord,
-  ControlledTextField,
-  ControlledNumberField,
-  ControlledBooleanField,
-  ControlledMultiOption,
-  ControlledMultiOptionSelect,
-  ControlledMultiOptionRadio,
-  ControlledTextSelectField,
-  ControlledNumberSelectField,
-  ControlledTextRadioField,
-  ControlledNumberRadioField,
-  FieldSetters,
-  FormControlAlgo,
-  FormOptionsValidate,
-  FormOptionsFieldProps,
-  HTMLOptions,
-  CreateFormOptions,
-  ControlledOptions,
-  CreateFormUncontrolledOptions,
-  UncontrolledOptions,
-  FormFields,
-  FormContextErrors,
-  FormContextFiles,
-  FormContextToJSON,
-  FormContext,
-  CreateFormProviderProps,
-  CreateForm,
-  CreateFormUncontrolled,
-  FormContextUncontrolled,
-  UncontrolledTextField,
-  UncontrolledNumberField,
-  UncontrolledBooleanField,
-  FormUncontrolAlgo,
-  FormFieldsUncontrolled,
-} from 'types';
+export * from './hooksControlled';
+export * from './hooksUncontrolled';
